move sync from local to remote at the end of playbook build
