+++ bob/build.go
@@ -8,9 +8,6 @@ import (
 	"github.com/benchkram/errz"
 
 	"github.com/benchkram/bob/bob/playbook"
-	"github.com/benchkram/bob/bobtask/hash"
-	"github.com/benchkram/bob/pkg/boblog"
-	"github.com/benchkram/bob/pkg/store"
 )
 
 var (
@@ -34,67 +31,18 @@ func (b *B) Build(ctx context.Context, taskName string) (err error) {
 	// Hint: Hash computation (playbook execution) can only start after
 	// nix dependencies are resolved.
 	// Nix dependencies are considered in the input hash of a task.
-	remote := ag.Remotestore()
-
-	playbook, err := ag.Playbook(
+	p, err := ag.Playbook(
 		taskName,
 		playbook.WithCachingEnabled(b.enableCaching),
 		playbook.WithPredictedNumOfTasks(len(ag.BTasks)),
 		playbook.WithMaxParallel(b.maxParallel),
-		playbook.WithRemoteStore(remote),
+		playbook.WithRemoteStore(ag.Remotestore()),
 		playbook.WithLocalStore(b.local),
 	)
 	errz.Fatal(err)
 
-	err = playbook.Build(ctx)
+	err = p.Build(ctx)
 	errz.Fatal(err)
 
-	if b.enableCaching && remote != nil {
-		// sync any newly generated artifacts with the remote store
-		syncFromLocalToRemote(ctx, b.local, remote, getArtifactIds(playbook, true))
-	}
-
 	return nil
-}
-
-// getArtifactIds returns the artifact ids of the given playbook (and optionally checks if the target exists first)
-func getArtifactIds(pbook *playbook.Playbook, checkForTarget bool) []hash.In {
-	artifactIds := []hash.In{}
-	for _, t := range pbook.Tasks {
-		if checkForTarget && !t.TargetExists() {
-			continue
-		}
-
-		h, err := t.HashIn()
-		if err != nil {
-			continue
-		}
-
-		artifactIds = append(artifactIds, h)
-	}
-	return artifactIds
-}
-
-// syncFromLocalToRemote syncs the artifacts from the local store to the remote store.
-func syncFromLocalToRemote(ctx context.Context, local store.Store, remote store.Store, artifactIds []hash.In) {
-	for _, a := range artifactIds {
-		err := store.Sync(ctx, local, remote, a.String())
-		if errors.Is(err, store.ErrArtifactAlreadyExists) {
-			boblog.Log.V(1).Info(fmt.Sprintf("artifact already exists on the remote [artifactId: %s]. skipping...", a.String()))
-			continue
-		} else if err != nil {
-			boblog.Log.V(1).Error(err, fmt.Sprintf("failed to sync from local to remote [artifactId: %s]", a.String()))
-			continue
-		}
-
-		// wait for the remote store to finish uploading this artifact. can be moved outside of the for loop but then
-		// we don't know which artifacts failed to upload.
-		err = remote.Done()
-		if err != nil {
-			boblog.Log.V(1).Error(err, fmt.Sprintf("failed to sync from local to remote [artifactId: %s]", a.String()))
-			continue
-		}
-
-		boblog.Log.V(1).Info(fmt.Sprintf("synced from local to remote [artifactId: %s]", a.String()))
-	}
 }
