+++ bob/playbook/build.go
@@ -8,7 +8,9 @@ import (
 	"sync"
 
 	"github.com/benchkram/bob/bobtask"
+	"github.com/benchkram/bob/bobtask/hash"
 	"github.com/benchkram/bob/pkg/boblog"
+	"github.com/benchkram/bob/pkg/store"
 )
 
 // Build the playbook starting at root.
@@ -92,6 +94,11 @@ func (p *Playbook) Build(ctx context.Context) (err error) {
 
 	p.summary(processedTasks)
 
+	if p.enableCaching && p.remoteStore != nil && p.localStore != nil {
+		// sync any newly generated artifacts with the remote store
+		syncFromLocalToRemote(ctx, p.localStore, p.remoteStore, getArtifactIds(p, true))
+	}
+
 	if len(processingErrors) > 0 {
 		// Pass only the very first processing error.
 		return processingErrors[0]
@@ -122,4 +129,46 @@ func logSkippedInputs(count int, taskname string, skippedInputs []string) int {
 	}
 
 	return count
+}
+
+// getArtifactIds returns the artifact ids of the given playbook (and optionally checks if the target exists first)
+func getArtifactIds(pbook *Playbook, checkForTarget bool) []hash.In {
+	var artifactIds []hash.In
+	for _, t := range pbook.Tasks {
+		if checkForTarget && !t.TargetExists() {
+			continue
+		}
+
+		h, err := t.HashIn()
+		if err != nil {
+			continue
+		}
+
+		artifactIds = append(artifactIds, h)
+	}
+	return artifactIds
+}
+
+// syncFromLocalToRemote syncs the artifacts from the local store to the remote store.
+func syncFromLocalToRemote(ctx context.Context, local store.Store, remote store.Store, artifactIds []hash.In) {
+	for _, a := range artifactIds {
+		err := store.Sync(ctx, local, remote, a.String())
+		if errors.Is(err, store.ErrArtifactAlreadyExists) {
+			boblog.Log.V(1).Info(fmt.Sprintf("artifact already exists on the remote [artifactId: %s]. skipping...", a.String()))
+			continue
+		} else if err != nil {
+			boblog.Log.V(1).Error(err, fmt.Sprintf("failed to sync from local to remote [artifactId: %s]", a.String()))
+			continue
+		}
+
+		// wait for the remote store to finish uploading this artifact. can be moved outside of the for loop but then
+		// we don't know which artifacts failed to upload.
+		err = remote.Done()
+		if err != nil {
+			boblog.Log.V(1).Error(err, fmt.Sprintf("failed to sync from local to remote [artifactId: %s]", a.String()))
+			continue
+		}
+
+		boblog.Log.V(1).Info(fmt.Sprintf("synced from local to remote [artifactId: %s]", a.String()))
+	}
 }
